изменила страницы регистрации, логина
