+++ src/entities/user/ui/user-info/unauthorized-user/unauthorized-user.tsx
@@ -1,31 +1,18 @@
-import classnames from 'classnames';
-import { useNavigate } from 'react-router-dom';
-
 import { Button } from 'shared/ui/button';
-
+import { VkIcon } from 'shared/ui/icons/vk-icon';
+import { handleRedirectVK } from 'shared/libs/utils';
 import styles from '../styles.module.css';
 
 export const UnauthorizedUser = () => {
-  const navigate = useNavigate();
-
   return (
     <div className={styles.buttonsWrapper}>
       <Button
-        extClassName={classnames(
-          styles['nested-buttons'],
-          styles['registration-button']
-        )}
         buttonType="primary"
-        label="Зарегистрироваться"
-        onClick={() => navigate('/register')}
+        actionType="submit"
+        customIcon={<VkIcon color="white" size="24" />}
+        label="Войти через ВКонтакте"
         size="extraLarge"
-      />
-      <Button
-        extClassName={styles['nested-buttons']}
-        buttonType="secondary"
-        label="Войти"
-        onClick={() => navigate('/login')}
-        size="extraLarge"
+        onClick={() => handleRedirectVK()}
       />
     </div>
   );
