+++ src/pages/vk-auth/index.tsx
@@ -27,22 +27,8 @@ export const VKAuthPage = () => {
 
         if (response.ok) {
           const userData = await response.json(); // Извлекаем данные из ответа
-          /*данные которые возвращаются при запросе access токена userData = {
-                    "access_token" : "token",
-                    "access_token_id" : "some_id",
-                    "user_id" : user_id,
-                    "phone": "",
-                    "phone_validated": 1592423373,
-                    "is_service": false,
-                    "email": "test-email@mail.ru",
-                    "source": 8,
-                    "source_description": "Auth by exchange token on change password"
-                }*/
-          //TODO: ХЗ какие данные придут с сервера, поэтому не могу продумать полноценно логики. Накинула примерный план
-          // если есть user в базе c vk_id === user_id, то этот пользователь уже есть
-          // --- заправшиваем данные пользователя из базы, устанавливаем в хранилище, устанавливаем токен, перенаправляем на главную
-          // иначе
-          // --- берем данные, которые вернул ВК, перенаправляем на форму регистрации для подтверждения или корректировки данных пользователя,
+
+          //TODO: Логика перенаправления на страницу регистрации, или вход
         } else {
           // Если ответ от сервера не успешный
           setIsError(true); // Обрабатываем ошибку
